+++ shared/types/common/game.ts
@@ -1,0 +1,5 @@
+import type { AvalonGameConfig } from '../games/avalon/options';
+
+export type TGameType = 'avalon';
+
+export type TGameConfig = AvalonGameConfig;