+++ shared/types/common/index.ts
@@ -1,0 +1,14 @@
+export type { TGameType, TGameConfig } from './game';
+export type {
+  TRoomState,
+  TVoteTarget,
+  VoteRoomResult,
+  VoteInRoom,
+  BaseRoomState,
+  CreatedRoomState,
+  LockedRoomState,
+  StartedRoomState,
+  RoomPlayer,
+  VoteOfPlayer,
+} from './room';
+export type { TRoomsList, TRoomInfo } from './room-list';