+++ shared/types/games/avalon/index.ts
@@ -1,0 +1,45 @@
+export type {
+  TRoles,
+  TGoodRoles,
+  TEvilRoles,
+  TUnknownRoles,
+  TGoodOptionalRoles,
+  TGoodRequiredRoles,
+  TEvilOptionalRoles,
+  TEvilRequiredRoles,
+  TOptionalRoles,
+  TRequiredRoles,
+  TLoyalty,
+  TVisibleRole,
+} from './roles';
+
+export type { Player, PlayerFeatures } from './player';
+export type { TMissionResult, MissionSettings, MissionWithResult } from './mission';
+export type { TVoteOption, TeamMember, PreVoteData } from './vote';
+export type { GameOptionsRoles, AvalonGameConfig, GameOptions } from './options';
+export type { PlayersSettings, GameRoles, GameSettings, GameSettingsWithRoles } from './settings';
+export type { AvalonGameState, VisualGameState, GameResults, TGameStage, TGameEndReasons } from './state';
+export type { TAssassinateType, TAssassinateResult, AssassinAddonData, AssassinateProgressData } from './addons';
+
+export type {
+  THistoryResults,
+  THistoryType,
+  THistoryStage,
+  HistoryBase,
+  HiddenHistory,
+  IAction,
+  IActionWithResult,
+  Vote,
+  AnonymousVoteResult,
+  PreVote,
+  HistoryVoteBase,
+  AnonymousHistoryVote,
+  HistoryVote,
+  THistoryVote,
+  HistoryMissionBase,
+  HistoryMission,
+  HistoryMissionHidden,
+  THistoryMission,
+  SwitchResult,
+  HistoryAssassinate,
+} from './history';